pyproject.toml
[project]
name = "fuelbillextraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "dotenv>=0.9.9",
    "flask>=3.1.1",
    "google>=3.0.0",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
]
